requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
base58==2.1.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
Cython==0.29.27
cytoolz==0.11.2
discord==1.7.3
discord.py==1.7.3
ecdsa==0.17.0
eth-hash==0.3.2
eth-keys==0.4.0
eth-typing==3.0.0
eth-utils==2.0.0
idna==3.3
more-itertools==8.12.0
multidict==6.0.2
py-bip39-bindings==0.1.8
py-ed25519-bindings==1.0.1
py-sr25519-bindings==0.1.4
pycryptodome==3.14.0
python-dotenv==0.19.2
requests==2.27.1
scalecodec==1.0.30
six==1.16.0
substrate-interface==1.1.8
toolz==0.11.2
typing-extensions==4.0.1
urllib3==1.26.8
websocket-client==1.2.3
xxhash==2.0.2
yarl==1.7.2
